controller: Enable WiFi AP mode on Core 0
Seems like, to access the SPI Flash, the ESP32 requires doing a
cross-core function call to disable the SPI cache on the other core.
This requires interrupts. However, interrupts are disabled on Core 1.

By reordering some stuff (initializing part of the flash and WiFi before
starting the core 1 task), and disabling the WiFi PHY calibration
storage (??), we can make it work and get into Soft-AP mode.

In the future we might have to make it so that we can re-enable
interrupts on core 1 or otherwise signal it, if we need to write to the
NVS key/value store (e.g. when changing config in the Web UI).