Remove 'detect game over before game started' logic
This was designed to detect a game over if the controller restarted
while a game was in progress. However, this can sometimes lead to double
uploads (if the flag goes 0 -> 1 -> 2 and we detect that),
and we don't need the "restart if wifi is down" logic now that we
back up scores to flash if we can't upload them.